Zola Migration
This marks the beginning of migrating the app over to Zola.

I am doing this cause Zola allows more freedom and a templating language i am familiar with.

The site will be rebuild and reordered so that users sort by not only sourcebook, but class, subclass and other things as well